+++ root/modules/user_pending/Controller.ts
@@ -58,17 +58,54 @@ export const User_Pending_Controller = {
 
   async DeleteUserPending(request: FastifyRequest, reply: FastifyReply) {
     const { id } = request.params as { id: number };
+    const token = request.cookies.refreshToken as string;
+
+    if (!token)
+      return reply.status(401).send({ message: "token não encontrado! " });
+
+    const decode = request.server.jwt.verify(token) as {
+      IDcompany: number;
+      role: string;
+    };
+
+    if (decode.role !== "admin")
+      return reply.status(403).send({ message: "Acesso negado!" });
 
     return await ServicesAcount.DeleteAcount(id);
   },
 
   async FindAllUserPending(request: FastifyRequest, reply: FastifyReply) {
+    const token = request.cookies.refreshToken as string;
+
+    if (!token)
+      return reply.status(401).send({ message: "token não encontrado! " });
+
+    const decode = request.server.jwt.verify(token) as {
+      IDcompany: number;
+      role: string;
+    };
+
+    if (decode.role !== "admin")
+      return reply.status(403).send({ message: "Acesso negado!" });
+
     const data = await ServicesAcount.FindAllAcount();
     return data;
   },
 
   async FindByIdUserPending(request: FastifyRequest, reply: FastifyReply) {
     const { id } = request.params as { id: number };
+    const token = request.cookies.refreshToken as string;
+
+    if (!token)
+      return reply.status(401).send({ message: "token não encontrado! " });
+
+    const decode = request.server.jwt.verify(token) as {
+      IDcompany: number;
+      role: string;
+    };
+
+    if (decode.role !== "admin")
+      return reply.status(403).send({ message: "Acesso negado!" });
 
     return ServicesAcount.FindByIdAcount(id);
   },
